+++ Source/SwiftLintFramework/Rules/Lint/PrivateSubjectRuleExamples.swift
@@ -1,4 +1,4 @@
-internal struct PrivatePassthroughSubjectRuleExamples {
+internal struct PrivateSubjectRuleExamples {
     static let nonTriggeringExamples: [Example] = [
         Example(
             #"""
@@ -25,6 +25,34 @@ internal struct PrivatePassthroughSubjectRuleExamples {
             #"""
             final class Foobar {
                 private let goodSubject: PassthroughSubject<Bool, Never> = .ini()
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private let goodSubject = CurrentValueSubject<Bool, Never>(false)
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private let goodSubject: CurrentValueSubject<Bool, Never>
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                fileprivate let goodSubject: CurrentValueSubject<String, Never>
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private let goodSubject: CurrentValueSubject<String, Never> = .ini("toto")
             }
             """#
         )
@@ -82,6 +110,60 @@ internal struct PrivatePassthroughSubjectRuleExamples {
                 private(set) let ↓badSubject = PassthroughSubject<Bool, Never>()
                 private let goodSubject: PassthroughSubject<Bool, Never>
                 private(set) let ↓anotherBadSubject = PassthroughSubject<Bool, Never>()
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                let ↓badSubject = CurrentValueSubject<Bool, Never>(true)
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                let ↓badSubject: CurrentValueSubject<Bool, Never>
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private(set) let ↓badSubject: CurrentValueSubject<Bool, Never>
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private(set) let ↓badSubject = CurrentValueSubject<Bool, Never>(false)
+            }
+            """#
+        )
+        ,
+        Example(
+            #"""
+            final class Foobar {
+                let goodSubject: CurrentValueSubject<String, Never> = .ini("toto")
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private let goodSubject: CurrentValueSubject<Bool, Never>
+                private(set) let ↓badSubject = CurrentValueSubject<Bool, Never>(false)
+                private(set) let ↓anotherBadSubject = CurrentValueSubject<Bool, Never>(false)
+            }
+            """#
+        ),
+        Example(
+            #"""
+            final class Foobar {
+                private(set) let ↓badSubject = CurrentValueSubject<Bool, Never>(false)
+                private let goodSubject: CurrentValueSubject<Bool, Never>
+                private(set) let ↓anotherBadSubject = CurrentValueSubject<Bool, Never>(true)
             }
             """#
         )