+++ Tests/LinuxMain.swift
@@ -1189,8 +1189,8 @@ extension PrivateOverFilePrivateRuleTests {
     ]
 }
 
-extension PrivatePassthroughSubjectRuleTests {
-    static var allTests: [(String, (PrivatePassthroughSubjectRuleTests) -> () throws -> Void)] = [
+extension PrivateSubjectRuleTests {
+    static var allTests: [(String, (PrivateSubjectRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
     ]
 }
@@ -1923,7 +1923,7 @@ XCTMain([
     testCase(PrivateActionRuleTests.allTests),
     testCase(PrivateOutletRuleTests.allTests),
     testCase(PrivateOverFilePrivateRuleTests.allTests),
-    testCase(PrivatePassthroughSubjectRuleTests.allTests),
+    testCase(PrivateSubjectRuleTests.allTests),
     testCase(PrivateUnitTestRuleTests.allTests),
     testCase(ProhibitedInterfaceBuilderRuleTests.allTests),
     testCase(ProhibitedSuperRuleTests.allTests),